Moving test steps and services for API tests to the same assemblies as E2E tests so they can share infrastructure
