+++ test/AdaptiveRemote.Backend.ApiTests/reqnroll.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://schemas.reqnroll.net/reqnroll-config-latest.json",
+
+  "bindingAssemblies": [
+    {
+      "assembly": "AdaptiveRemote.EndToEndTests.Steps"
+    }
+  ]
+}
