+++ test/AdaptiveRemote.EndToEndTests.Steps/Backend/CommonSteps.cs
@@ -1,11 +1,9 @@
-using System.Net;
-using System.Text.Json;
-using AdaptiveRemote.Backend.ApiTests.Support;
 using AdaptiveRemote.Contracts;
+using AdaptiveRemote.EndToEndTests.TestServices.Backend;
 using FluentAssertions;
 using Reqnroll;
 
-namespace AdaptiveRemote.Backend.ApiTests.StepDefinitions;
+namespace AdaptiveRemote.EndToEndTests.Steps.Backend;
 
 [Binding]
 public class CommonSteps : IDisposable
@@ -22,7 +20,7 @@ public class CommonSteps : IDisposable
         => new(_context.Fixture.ServiceUrl);
 
     [Given(@"CompiledLayoutService is running")]
-    public async Task GivenCompiledLayoutServiceIsRunning()
+    public async Task GivenCompiledLayoutServiceIsRunningAsync()
     {
         await _context.Fixture.StartServiceAsync();
     }
