+++ test/AdaptiveRemote.EndtoEndTests.TestServices/AdaptiveRemote.EndtoEndTests.TestServices.csproj
@@ -8,12 +8,15 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AWSSDK.DynamoDBv2" />
+    <PackageReference Include="AWSSDK.SQS" />
     <PackageReference Include="Deque.AxeCore.Commons" />
     <PackageReference Include="Deque.AxeCore.Playwright" />
     <PackageReference Include="FluentAssertions" />
+    <PackageReference Include="Microsoft.Playwright" />
     <PackageReference Include="MSTest.TestFramework" />
     <PackageReference Include="StreamJsonRpc" />
-    <PackageReference Include="Microsoft.Playwright" />
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" />
   </ItemGroup>
 
   <ItemGroup>
